Remove unneeded files
In the people code-along, developers are supposed to interact with
`people.csv`. They do not interact directly with any of the text files
stored in `census`. They also don't run the `people_generator.rb`
script, which is what was originally used to create `people.csv`.

Those text files and the generator scripts are just plumbing that was
used to create the lesson. Once we have `people.csv`, we don't need them
anymore. They also made references to gender which is a sensitive topic.

This commit removes them. Close #39.