Update tests to reflect changes to data
Rerunning the `generate_people.rb` script to remove gender changed some
of the data in `people.csv`, probably because the version of
the text files in `census` that were used to create `people.csv`
originally are different from what's in `census` now.

This commit updates the test so that they're looking for the correct
average data for this version of `people.csv`.